requirements.txt
streamlit 

nest_asyncio

aiogram
python-telegram-bot
